Disable filtering of available packages for availability of their depexts
OPAM by default filters out packages for which the system on which it is
running does not have the depexts. Which is nice in theory but when
building a lockfile we want all packages to be available.

The problem here is that some configurations do not have a queryable
package list, e.g. when using Docker containers the package list might
have been deleted to save on space in the container, thus
`available_packages` would be just a subset and missing packages,
despite these existing in the OPAM repositories.