Read package flags from OPAM metadata
This changes the logic from hardcoding the compiler "base packages" to
instead read the OPAM flags to check whether the package in question is
a compiler in which case it will be considered a base package.

It also adds code to check for packages whether they are `conf` packages,
which could be used to check for virtual packages.

Inspired by #327, this would allow `dkml-base-compiler` to be considered
a base package and not accidentally vendor it. However it doesn't yet
allow using the DKML compiler, because the name of the `ocaml` conf
package is still hardcoded with no way to override it (yet).