Fix sort order of avoid-versions
Previously the solver discarded everything in the back of the list, thus
not only avoiding the avoid-version packages but effectively removing
them from the candidate set. Now we keep the groups of packages
consecutive which makes the solver pick them if there is no other
option, while still trying to avoid them if other options are available.