Error when multiple dev-repos collide on dir
The duniverse directory of a package is implied by its dev-repo. It's
possible for different dev-repos to map to the same directory. Prior to
this change, one lucky package would be downloaded to the expected place
and the others would be silently ignored. More precisely, each package
would be downloaded and placed in the duniverse dir implied by their
dev-repo, but first any existing duniverse dir of the same name would be
removed.

After this change, the same situation results in an error when running
`opam monorepo lock`.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>