Simplify error message when deps don't build with dune
When trying to generate a lockfile on a package which has dependencies
which don't use dune as their build system, the diagnostic message
generated by opam-0install is of little use. This change removes the
opam-0install diagnostic message from the output of `opam monorepo lock`
in the case where it fails due to dependencies not building with dune.

Additionally this rewords the error message to highlight the salient
information. In particular the list of non-dune packages is printed as a
bulleted list rather than a comma-separated list inside the error
message (it often takes me a minute to find the list of packages in the
old error message).