Remove custom deduplication of packages and use opam solver instead
Convert the constraint "if X and Y are from the same repository" into
a conflict relationship in the opam metadata between `X` and all the
version of `Y` which do not have the same exact version as `X`.

This allow to remove the adhoc post-processing of opam solution which
was going very wrong in lots of case. Hopefully, the results will now
be much closer to what you would expect `opam` to produce.